android: removed Android app
The planned Android app was removed because I soon
will ditch Android and no longer use it. (May 2019)
Android has become too bloated and too restricted
over the past years in terms of user freedom.

I'm looking forward to buy the upcoming
 > "Purism Librem 5" Phone,
which respects privacy by default and has hardware
kill switches.

I'm focusing on a touch-friendly version of the Qt
version which is optimized for the small screen of
the Librem 5.